# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wild Haven employee operations API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Staff authentication and profile
  - name: dashboard
    description: Mission and donation overview
  - name: donors
    description: Donor records
  - name: donations
    description: Donation records
  - name: campaigns
    description: Conservation campaigns
  - name: impact
    description: Impact reporting
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/me:
    get:
      operationId: getCurrentStaff
      tags: [auth]
      summary: Get the current staff member
      responses:
        "200":
          description: Current staff profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get the operations dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /activity:
    get:
      operationId: listActivity
      tags: [dashboard]
      summary: List recent staff activity
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: List searchable donor records
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Donor records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donor"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createDonor
      tags: [donors]
      summary: Create a donor record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorInput"
      responses:
        "201":
          description: Donor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteAllDonors
      tags: [donors]
      summary: Not supported; included only to keep the collection contract explicit
      responses:
        "405":
          description: Method not allowed
  /donors/{id}:
    get:
      operationId: getDonor
      tags: [donors]
      summary: Get a donor record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Donor record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateDonor
      tags: [donors]
      summary: Update a donor record
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorUpdate"
      responses:
        "200":
          description: Donor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDonor
      tags: [donors]
      summary: Delete a donor record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Donor deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donation records
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DonationStatus"
        - name: campaignId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Donation records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Record a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Donation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get a donation record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Donation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateDonation
      tags: [donations]
      summary: Update a donation record
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationUpdate"
      responses:
        "200":
          description: Donation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDonation
      tags: [donations]
      summary: Delete a donation record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Donation deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List conservation campaigns
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CampaignStatus"
      responses:
        "200":
          description: Campaign records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a conservation campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a conservation campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Campaign record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a conservation campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a conservation campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Campaign deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /impact:
    get:
      operationId: getImpactSummary
      tags: [impact]
      summary: Get grouped conservation impact metrics
      responses:
        "200":
          description: Impact summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 100
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Record not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    StaffMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/StaffRole"
        initials:
          type: string
      required: [id, name, email, role, initials]
    StaffRole:
      type: string
      enum: [admin, manager, coordinator]
    DashboardSummary:
      type: object
      properties:
        totalRaised:
          type: number
        totalRaisedChange:
          type: number
        donorCount:
          type: integer
        donorCountChange:
          type: number
        activeCampaignCount:
          type: integer
        monthlyRecurring:
          type: number
        monthlyRecurringChange:
          type: number
        campaignProgress:
          type: array
          items:
            $ref: "#/components/schemas/CampaignProgress"
        donationTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
      required:
        - totalRaised
        - totalRaisedChange
        - donorCount
        - donorCountChange
        - activeCampaignCount
        - monthlyRecurring
        - monthlyRecurringChange
        - campaignProgress
        - donationTrend
    CampaignProgress:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        raised:
          type: number
        goal:
          type: number
        percent:
          type: number
        status:
          $ref: "#/components/schemas/CampaignStatus"
      required: [id, name, raised, goal, percent, status]
    TrendPoint:
      type: object
      properties:
        label:
          type: string
        value:
          type: number
      required: [label, value]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [donation, donor, campaign, report]
        title:
          type: string
        detail:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, detail, createdAt]
    Donor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        totalGiven:
          type: number
        donationCount:
          type: integer
        lastGiftAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, phone, city, totalGiven, donationCount, lastGiftAt, createdAt]
    DonorInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
      required: [name, email]
    DonorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
    Donation:
      type: object
      properties:
        id:
          type: integer
        donorId:
          type: integer
        donorName:
          type: string
        amount:
          type: number
        currency:
          type: string
        frequency:
          $ref: "#/components/schemas/DonationFrequency"
        status:
          $ref: "#/components/schemas/DonationStatus"
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        donatedAt:
          type: string
          format: date-time
        receiptNumber:
          type: string
      required: [id, donorId, donorName, amount, currency, frequency, status, campaignId, campaignName, donatedAt, receiptNumber]
    DonationInput:
      type: object
      properties:
        donorId:
          type: integer
          minimum: 1
        amount:
          type: number
          exclusiveMinimum: 0
        frequency:
          $ref: "#/components/schemas/DonationFrequency"
        status:
          $ref: "#/components/schemas/DonationStatus"
        campaignId:
          type: ["integer", "null"]
        donatedAt:
          type: string
          format: date-time
      required: [donorId, amount, frequency, status]
    DonationUpdate:
      type: object
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
        frequency:
          $ref: "#/components/schemas/DonationFrequency"
        status:
          $ref: "#/components/schemas/DonationStatus"
        campaignId:
          type: ["integer", "null"]
        donatedAt:
          type: string
          format: date-time
    DonationFrequency:
      type: string
      enum: [one_time, monthly]
    DonationStatus:
      type: string
      enum: [completed, pending, refunded]
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        species:
          type: string
        location:
          type: string
        description:
          type: string
        goal:
          type: number
        raised:
          type: number
        status:
          $ref: "#/components/schemas/CampaignStatus"
        supporters:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, species, location, description, goal, raised, status, supporters, createdAt]
    CampaignInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        species:
          type: string
        location:
          type: string
        description:
          type: string
        goal:
          type: number
          exclusiveMinimum: 0
        status:
          $ref: "#/components/schemas/CampaignStatus"
      required: [name, species, location, description, goal, status]
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        species:
          type: string
        location:
          type: string
        description:
          type: string
        goal:
          type: number
          exclusiveMinimum: 0
        status:
          $ref: "#/components/schemas/CampaignStatus"
    CampaignStatus:
      type: string
      enum: [active, paused, completed]
    ImpactSummary:
      type: object
      properties:
        acresProtected:
          type: number
        rescueMissions:
          type: integer
        activeGuards:
          type: integer
        communityPrograms:
          type: integer
        allocation:
          type: array
          items:
            $ref: "#/components/schemas/ImpactAllocation"
      required: [acresProtected, rescueMissions, activeGuards, communityPrograms, allocation]
    ImpactAllocation:
      type: object
      properties:
        label:
          type: string
        amount:
          type: number
        percent:
          type: number
      required: [label, amount, percent]