# lib/api-zod/src/generated/types/campaignStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wild Haven employee operations API
 * OpenAPI spec version: 0.1.0
 */

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;