# lib/api-zod/src/generated/types/campaignUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wild Haven employee operations API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from './campaignStatus';

export interface CampaignUpdate {
  /** @minLength 2 */
  name?: string;
  species?: string;
  location?: string;
  description?: string;
  /** @exclusiveMinimum 0 */
  goal?: number;
  status?: CampaignStatus;
}