# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wild Haven employee operations API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationFrequency } from './donationFrequency';
import type { DonationStatus } from './donationStatus';

export interface Donation {
  id: number;
  donorId: number;
  donorName: string;
  amount: number;
  currency: string;
  frequency: DonationFrequency;
  status: DonationStatus;
  /** @nullable */
  campaignId: number | null;
  /** @nullable */
  campaignName: string | null;
  donatedAt: Date;
  receiptNumber: string;
}