# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wild Haven employee operations API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationFrequency } from './donationFrequency';
import type { DonationStatus } from './donationStatus';

export interface DonationInput {
  /** @minimum 1 */
  donorId: number;
  /** @exclusiveMinimum 0 */
  amount: number;
  frequency: DonationFrequency;
  status: DonationStatus;
  /** @nullable */
  campaignId?: number | null;
  donatedAt?: Date;
}