# lib/api-zod/src/generated/types/impactAllocation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wild Haven employee operations API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpactAllocation {
  label: string;
  amount: number;
  percent: number;
}